fix: address PR feedback and CI failures
- Revert ty pin to v0.0.17 to match main's CI; align pre-commit hook to
  the same version. Remove the now-unneeded ty:ignore on external_id.
- Make form-error and help-text assertions more specific (CodeQL
  py/incomplete-url-substring-sanitization on the previous bare-domain
  substring matches).
- Split TestEmailChannelForm and TestEmailDomainAllowlist out of
  test_email_channel.py into focused modules to address CodeScene
  Lines-of-Code.
