+++ .pre-commit-config.yaml
@@ -53,8 +53,8 @@ repos:
         files: ^components/chat_widget/src/.*\.(ts|tsx|css)$
         types: [ file ]
   - repo: https://github.com/allganize/ty-pre-commit
-    # Ty version.
-    rev: v0.0.32
+    # Ty version. Keep in sync with the `ty` entry in pyproject.toml's [dependency-groups.dev].
+    rev: v0.0.17
     hooks:
       # Run the type checker.
       - id: ty-check
