+++ apps/channels/channels_v2/email_channel.py
@@ -239,8 +239,7 @@ class EmailChannel(ChannelBase):
         if self.experiment_session and self._sender_instance:
             msg_id = self._sender_instance.last_message_id
             if msg_id and not _has_email_message_id(self.experiment_session.external_id):
-                # ty 0.0.32 false positive: external_id is a CharField that accepts str at runtime.
-                self.experiment_session.external_id = msg_id  # ty: ignore[invalid-assignment]
+                self.experiment_session.external_id = msg_id
                 try:
                     self.experiment_session.save(update_fields=["external_id"])
                 except IntegrityError:
