+++ apps/channels/tests/test_email_channel.py
@@ -15,10 +15,8 @@ from apps.channels.channels_v2.email_channel import (
     get_email_experiment_channel,
 )
 from apps.channels.datamodels import EmailMessage
-from apps.channels.forms import EmailChannelForm
 from apps.channels.models import ChannelPlatform, ExperimentChannel
 from apps.channels.tasks import handle_email_message
-from apps.channels.utils import get_allowed_email_domains, is_email_domain_allowed
 from apps.chat.channels import MESSAGE_TYPES
 from apps.chat.models import Chat
 from apps.experiments.models import ExperimentSession, Participant
@@ -140,188 +138,6 @@ class TestEmailMessageParse:
         inbound = _make_inbound_message(text="Just a simple message")
         result = EmailMessage.parse(inbound)
         assert result.message_text == "Just a simple message"
-
-
-class TestEmailDomainAllowlist:
-    @override_settings(EMAIL_CHANNEL_ALLOWED_DOMAINS=["example.com", "*.foo.com"])
-    def test_exact_match_allowed(self):
-        assert is_email_domain_allowed("user@example.com") is True
-
-    @override_settings(EMAIL_CHANNEL_ALLOWED_DOMAINS=["example.com", "*.foo.com"])
-    def test_wildcard_match_allowed(self):
-        assert is_email_domain_allowed("user@mail.foo.com") is True
-
-    @override_settings(EMAIL_CHANNEL_ALLOWED_DOMAINS=["example.com", "*.foo.com"])
-    def test_bare_domain_does_not_match_wildcard(self):
-        # *.foo.com matches subdomains, not the bare domain.
-        assert is_email_domain_allowed("user@foo.com") is False
-
-    @override_settings(EMAIL_CHANNEL_ALLOWED_DOMAINS=["example.com"])
-    def test_disallowed_domain_rejected(self):
-        assert is_email_domain_allowed("user@bar.com") is False
-
-    @override_settings(EMAIL_CHANNEL_ALLOWED_DOMAINS=[])
-    def test_empty_setting_rejects_everything(self):
-        assert is_email_domain_allowed("user@example.com") is False
-
-    @override_settings(EMAIL_CHANNEL_ALLOWED_DOMAINS=["example.com"])
-    def test_malformed_address_rejected(self):
-        assert is_email_domain_allowed("not-an-email") is False
-        assert is_email_domain_allowed("") is False
-
-    @override_settings(EMAIL_CHANNEL_ALLOWED_DOMAINS=["example.com"])
-    def test_case_insensitive_match(self):
-        assert is_email_domain_allowed("user@Example.COM") is True
-
-    @override_settings(EMAIL_CHANNEL_ALLOWED_DOMAINS=["example.com", "*.foo.com"])
-    def test_get_allowed_email_domains_returns_list(self):
-        assert get_allowed_email_domains() == ["example.com", "*.foo.com"]
-
-    @override_settings(EMAIL_CHANNEL_ALLOWED_DOMAINS=[])
-    def test_get_allowed_email_domains_returns_empty_list_when_unset(self):
-        assert get_allowed_email_domains() == []
-
-
-@pytest.mark.django_db()
-class TestEmailChannelForm:
-    @override_settings(EMAIL_CHANNEL_ALLOWED_DOMAINS=["chat.openchatstudio.com"])
-    def test_valid_form(self, experiment):
-        form = EmailChannelForm(
-            experiment=experiment,
-            data={"email_address": "support@chat.openchatstudio.com", "platform": "email"},
-        )
-        assert form.is_valid(), form.errors
-
-    def test_email_address_required(self, experiment):
-        form = EmailChannelForm(experiment=experiment, data={"platform": "email"})
-        assert not form.is_valid()
-        assert "email_address" in form.errors
-
-    @override_settings(EMAIL_CHANNEL_ALLOWED_DOMAINS=["chat.openchatstudio.com"])
-    def test_from_address_optional(self, experiment):
-        form = EmailChannelForm(
-            experiment=experiment,
-            data={"email_address": "support@chat.openchatstudio.com", "platform": "email"},
-        )
-        assert form.is_valid()
-        assert form.cleaned_data.get("from_address", "") == ""
-
-    @override_settings(EMAIL_CHANNEL_ALLOWED_DOMAINS=["chat.openchatstudio.com"])
-    def test_is_default_defaults_to_false(self, experiment):
-        form = EmailChannelForm(
-            experiment=experiment,
-            data={"email_address": "support@chat.openchatstudio.com", "platform": "email"},
-        )
-        assert form.is_valid()
-        assert form.cleaned_data["is_default"] is False
-
-    def test_duplicate_default_rejected(self, experiment):
-        """Only one default email channel per team."""
-        _make_email_channel(experiment.team, email_address="first@chat.openchatstudio.com", is_default=True)
-
-        form = EmailChannelForm(
-            experiment=experiment,
-            data={
-                "email_address": "second@chat.openchatstudio.com",
-                "is_default": True,
-                "platform": "email",
-            },
-        )
-        assert not form.is_valid()
-        assert "is_default" in form.errors
-
-    @override_settings(EMAIL_CHANNEL_ALLOWED_DOMAINS=["chat.openchatstudio.com"])
-    def test_duplicate_default_allowed_when_editing_same_channel(self, experiment):
-        """Editing the existing default channel should not trigger uniqueness error."""
-        channel = _make_email_channel(
-            experiment.team, experiment=experiment, email_address="first@chat.openchatstudio.com", is_default=True
-        )
-
-        form = EmailChannelForm(
-            experiment=experiment,
-            channel=channel,
-            data={
-                "email_address": "first@chat.openchatstudio.com",
-                "is_default": True,
-                "platform": "email",
-            },
-        )
-        assert form.is_valid(), form.errors
-
-    @override_settings(EMAIL_CHANNEL_ALLOWED_DOMAINS=["chat.openchatstudio.com"])
-    def test_email_address_on_allowed_domain_accepted(self, experiment):
-        form = EmailChannelForm(
-            experiment=experiment,
-            data={"email_address": "support@chat.openchatstudio.com", "platform": "email"},
-        )
-        assert form.is_valid(), form.errors
-
-    @override_settings(EMAIL_CHANNEL_ALLOWED_DOMAINS=["*.openchatstudio.com"])
-    def test_email_address_on_allowed_wildcard_accepted(self, experiment):
-        form = EmailChannelForm(
-            experiment=experiment,
-            data={"email_address": "support@chat.openchatstudio.com", "platform": "email"},
-        )
-        assert form.is_valid(), form.errors
-
-    @override_settings(EMAIL_CHANNEL_ALLOWED_DOMAINS=["chat.openchatstudio.com"])
-    def test_email_address_on_disallowed_domain_rejected(self, experiment):
-        form = EmailChannelForm(
-            experiment=experiment,
-            data={"email_address": "support@evil.example.com", "platform": "email"},
-        )
-        assert not form.is_valid()
-        assert "email_address" in form.errors
-        # Error message should mention the allowed list so admins can self-serve.
-        assert "chat.openchatstudio.com" in str(form.errors["email_address"])
-
-    @override_settings(EMAIL_CHANNEL_ALLOWED_DOMAINS=[])
-    def test_email_address_rejected_when_setting_empty(self, experiment):
-        form = EmailChannelForm(
-            experiment=experiment,
-            data={"email_address": "support@chat.openchatstudio.com", "platform": "email"},
-        )
-        assert not form.is_valid()
-        assert "email_address" in form.errors
-        assert "no allowed domains" in str(form.errors["email_address"]).lower()
-
-    @override_settings(EMAIL_CHANNEL_ALLOWED_DOMAINS=["chat.openchatstudio.com"])
-    def test_from_address_validated_when_set(self, experiment):
-        form = EmailChannelForm(
-            experiment=experiment,
-            data={
-                "email_address": "support@chat.openchatstudio.com",
-                "from_address": "noreply@evil.example.com",
-                "platform": "email",
-            },
-        )
-        assert not form.is_valid()
-        assert "from_address" in form.errors
-
-    @override_settings(EMAIL_CHANNEL_ALLOWED_DOMAINS=["chat.openchatstudio.com"])
-    def test_from_address_skipped_when_blank(self, experiment):
-        form = EmailChannelForm(
-            experiment=experiment,
-            data={
-                "email_address": "support@chat.openchatstudio.com",
-                "from_address": "",
-                "platform": "email",
-            },
-        )
-        assert form.is_valid(), form.errors
-
-    @override_settings(EMAIL_CHANNEL_ALLOWED_DOMAINS=["chat.openchatstudio.com", "*.example.com"])
-    def test_help_text_lists_allowed_domains(self, experiment):
-        form = EmailChannelForm(experiment=experiment)
-        help_text = form.fields["email_address"].help_text
-        assert "chat.openchatstudio.com" in help_text
-        assert "*.example.com" in help_text
-
-    @override_settings(EMAIL_CHANNEL_ALLOWED_DOMAINS=[])
-    def test_help_text_warns_when_no_domains_configured(self, experiment):
-        form = EmailChannelForm(experiment=experiment)
-        help_text = form.fields["email_address"].help_text
-        assert "no allowed domains" in help_text.lower()
 
 
 @pytest.mark.django_db()
